requirements.txt
ipykernel
gdown
PyYAML    
python-box ==6.0.2   
joblib  
ensure==1.0.2 
pandas
seaborn
matplotlib
scikit-learn
tensorflow
mlflow
python-dotenv

-e .
